fix: fix build errors for `act-cli` and `oh-my-posh`
For `act`, we remove the patch as the current required go version is 1.24 and the oldest chroot (Fedora 42)
is on go 1.24.9. This is situation is bound to repeat itself if `act` decides to upgrade the required go
version before 42 goes EOL.

For `act-cli`, there was no simple solution, as the package requires go 1.25 and already uses code that only
compiles on that version. I have decided to drop the Fedora 42 chroot for this package.

A TODO is to expand our tests to run on all available chroots that we normally build for, but this is a pain.